pyproject.toml
[tool.poetry]
name = "django-docker-template"
version = "0.1.0"
description = ""
authors = ["BohdanSizomin"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.0"
djangorestframework = "^3.14.0"
django-debug-toolbar = "^4.2.0"
python-dotenv = "^1.0.0"
flake8 = "^6.1.0"
psycopg2-binary = "^2.9.9"
gunicorn = "^21.2.0"
whitenoise = "^6.6.0"
pre-commit = "^3.6.0"

[tool.poetry.group.dev.dependencies]
django-pytest = "^0.2.0"
black = "^23.12.0"
pytest-django = "^4.7.0"
pytest-xdist = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
